fix(api): patient route logic changes
refs. metriport/metriport-internal#2040
